ARUHA-657 Fix committing to subscription for closed stream
It is possible to have situation, when stream is closing, and at the same time there is a task for streaming data to client in subscription queue. In this case streaming session was closed without waiting for uncommitted events. This fix moves calculation of uncommitted events to the very end of StreamingState (so no further writes may occur after this calculation).